small header search bar
I've introduced a small search bar directly into the header navbar. It is non-functional (the current search function is the "mockSubmit" function). For some reason the search button and magnifying glass icon are not standing out against the black background, and the "color" variable isn't changing the color of the icon. More tinkering is required.