Save refreshed token in existing session, not a new one
When the updateSession function was calling
storage.getSession, it wasn't passing in the cookie. As a result,
it was always generating a _new_ orphan session, not bound to the
user's cookie.

The consequence of this bug was that if the tokens ever needed
refresh, the refresh token would be submitted on _every_ request
thereafter.

updateSession must take an existing session as a parameter.