Display just-submitted Circular in index
Add optimistic UI to the Circular index to include the
just-submitted Circular in the list.

The list of Circulars comes from OpenSearch, which is updated
asynchronously in the table streams Lambda. If a new Circular has
been submitted, add it to the list to provide visible feedback that
the form was processed.