addressed regex comments in PR#976
made changes requested by lpsinger in PR#976 wrt the circularRedirect function in circulars.server.ts. regex is more flexible: the leading and trailing whitespaces are optional, there is now a separate capturing group for the circular ID and optional non-capturing groups for gcn and circular (non-case sensitive). Also modified the circular ID capturing group to account for instances of negative and/or non-integer circular IDs. when a circular ID that doesn't exist is searched for, rather than throwing an error, it proceeds to a standard search.