shortened circularRedirect function
removed first conditional in favor of defining the circularId with parseFloat and then changing the conditional after to check if circularId is a number. In the parseFloat, had to add an empty string to not provide parseFloat undefined, which breaks TypeScript. Also altered the call to circularRedirect in index.tsx to only call when query is a string