[nasa-gcn/gcn.nasa.gov#943] redirect to circular page when searching specific ID (#976)
* redirect to circular page when searching specific ID

in circulars search, users can prepend their search with "circular:" followed by a specific circular number to redirect to the specific circular webpage

* redirect to specific circular is now more flexible

the redirect is now case insensitive and now triggers for instances where "gcn" or "circular" are used as well as the circular number on its own. should not conflict with any possible object keyword searches outside of perhaps Fermi trigger numbers without an accompanying non-numeric value.

* added circulars hint

used Hint from ~/components/Hint

* addressed regex comments in PR#976

made changes requested by lpsinger in PR#976 wrt the circularRedirect function in circulars.server.ts. regex is more flexible: the leading and trailing whitespaces are optional, there is now a separate capturing group for the circular ID and optional non-capturing groups for gcn and circular (non-case sensitive). Also modified the circular ID capturing group to account for instances of negative and/or non-integer circular IDs. when a circular ID that doesn't exist is searched for, rather than throwing an error, it proceeds to a standard search.

* added search hint item

* shortened circularRedirect function

removed first conditional in favor of defining the circularId with parseFloat and then changing the conditional after to check if circularId is a number. In the parseFloat, had to add an empty string to not provide parseFloat undefined, which breaks TypeScript. Also altered the call to circularRedirect in index.tsx to only call when query is a string

* Update app/routes/circulars/circulars.server.ts

Co-authored-by: Leo Singer <leo.p.singer@nasa.gov>

* Update app/routes/circulars/circulars.server.ts

Co-authored-by: Leo Singer <leo.p.singer@nasa.gov>

* fixed hint in circulars search

fixed issue with way search hint was used in index.tsx per Leo

* Update app/routes/circulars/index.tsx

Co-authored-by: Leo Singer <leo.p.singer@nasa.gov>

* Update app/routes/circulars/circulars.server.ts

Co-authored-by: Leo Singer <leo.p.singer@nasa.gov>

* Update app/routes/circulars/index.tsx

Co-authored-by: Leo Singer <leo.p.singer@nasa.gov>

* Update app/routes/circulars/index.tsx

Co-authored-by: Leo Singer <leo.p.singer@nasa.gov>

* fix prettier issue

---------

Co-authored-by: Leo Singer <leo.p.singer@nasa.gov>