added date range filter
using the useLocation method, it was possible to implement the correct onChange behaviour for the date range picker. I found that it is necessary to use a forward slash instead of a hyphen in order to preserve the date range filter when going back after navigating to a specific circular. The query input is also changed to remove the onChange, which removes the useState dependency and also alters the behaviour so a query request is only sent when the search button or the enter key is pressed. Prior to this, a query request would be sent with each character change as a user is typing their full search.