moved date range inside form
this makes it unnecessary (sic?) to use useLocation when defining url parameters inside default function. However, the current ui is definitely not final and will need to incorporate an as-yet-created custom expanding css element. There also needs to be a separate PR for a custom css element that incorporates both the prior element and a series of options, including expanding the date range selector. However, this seems to revert the fix Leo made to not populate the url with the empty date range parameters upon submitting a query. Will need to investigate further to fix.