implemented fuzzy date range in search function
conditional that ignores the last parameter if start or end dates are provided, the last parameter is ignored and the start and end Times are defined by dates parameters. If these are not provided, the last parameter defines the start and end times using the fuzzyTimeRange function. Next step is to implement this in index.tsx