made fuzzy date startDate
moved the fuzzy date to using the startDate parameter rather than the last parameter and then removed the last parameter. new function does a regex to check for anticipated formats for the date. If it does not recognize the string being in the date format, it will direct it to the fuzzy time range function. validated that both 2023-01-01 and ytd both produce the same number of results.