added buttons for fuzzy dates
the fuzzy date buttons will not work until #1172 is merged. I also found that there is an odd issue with the current implementation where one needs to select the specific fuzzy date option twice before it will actuall submit the url properly. likely has something to do with the way setDateRange is written.