Incrementally implement Astro Flavored Markdown
This is an idea for incremental adoption and development of Astro
Flavored Markdown. The idea is that as people are continuing to
submit plaintext GCN Circulars, we can parse _just_ the astronomy
related syntax but not the general Markdown syntax. So, for example,
whitespace in GCN Circualrs is treated the same as it is now, and
none of the Markdown syntax such as `# Headings` or `**emphasis**`
are processed, but GCN Circulars cross references _are_ processed.

The way this works is that we create a fake Markdown syntax tree
for the GCN Circular that puts all of its text in a single
paragraph, then we pass that through the Astro Flavored Markdown
processor.

This patch also demonstrates enriching Astro Flavored Markdown
with React components. We transform each Astro Flavored Markdown
node into an HTML [`<data>`](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/data)
element with the `class` attribute set to the Astro Flavored
Markdown type and the `value` attribute set to the data value.
The content is the original text.

For example, the following Markdown text:

```md
See GCN Circular 123a
```

is transformed to this Markdown AST:

```json
{
  "type": "root",
  "children": [
    {
      "type": "paragraph",
      "children": [
        {
          "type": "text",
          "value": "see GCN Circular"
        },
        {
          "type": "text",
          "value": "123a",
          "data": {
            "class": "gcn-circular",
            "value": 123.5
          }
        }
      ]
    }
  ]
}
```

And this, in turn, is transformed to the following HTML:

```html
<p>
  See GCN Circular
  <data class="gcn-circular" value="123.5">
    123a
  </data>
</p>
```

Finally, at the HTML to React stage (which is done by
[rehype-react](https://github.com/rehypejs/rehype-react)) we can
transform HTML `<data>` elements to React components of our choosing.