revert to dropdown button
updated dropdown button content to use the style developed for the advanced search filters (so they are now radio buttons and the custom date range div expands or collapses with the selection of the custom radio). However, the flex row and flex columns are no longer aligned correctly with this change, and the issue with the dropdown button label having an overflow persists.