fix fuzzy date to custom date switch
going from a fuzzy date to a custom date where just the end date is set no longer has the issue of the start date still being set to the fuzzy date. Also added labels to the date range selector so that instances of selecting just a start date or just an end date will look better.