hide dropdowns on click away
this initial commit adds the basic functionality for listening for clicking away from a dropdown, which will cause the dropdown to be hidden from the user unless the dropdown button is pressed again. However, this implementation seems to interfere with the dropdown button interaction slightly; if the menu is expanded, pressing the button again does not collapse the menu, though clicking anywhere else on the page will collapse the dropdown. This is not the case in the currently deployed commit, so it seems to be due to the implementation made in this pr.