initial commit
this is an initial commit that includes a small change to the styleguide page and creates a new page for the markdown primer. Further commits will add to the markdown primer page.

added links to markdown page

added relevant links and references to the markdown documentation page as well as an item in the navigation bar

use more precise language

URL hyperlinks are already supported via automatic conversion, so reference to this is removed.

hide page  behind feature flag

the markdown page will not load unless the feature flag is set, nor will the element in the sidebar. However, the two strings in the mdx pages that have been added still need to be excluded. Currently investigating how that can be accomplished. Barring that, can just remove them until such a time that the feature flag is removed.

revert markdown references

remove references to markdown primer in other pages until such a time that we enable the markdown feature

basic text formatting

header guide

run prettier

remove unnecessary import from styleguide

... and from index

added lists

add quotes and code

also add a section for sublists

added tables

also minor edits to a couple of the previous sections

added GCN cross reference

was not sure how much to edit the cross-reference syntax function as it already existed, so included it in its current form along with some explanatory text beforehand.

run prettier

apply changes

these changes are the ones made in commits following a merge