partial fix to switch from mobile to desktop transition
in this implementation, an expanded menu will persist when switching between a desktop and mobile window width, but at desktop widths, the overlay will not render over the page as it currently does. This behaviour could be changed such that any change from mobile to desktop widths would result in the expanded boolean being set to false. Outstanding issue: current implementation results in a warning about useLayoutEffect and differing behaviour on server vs client. This is because window is not defined on the server, so there are conditionals to prevent the window references from causing an error.