add conditional wildcard queries
if a user submits a query that consists of what could be an object identifier (eg `230919a`), the search function will automatically add wildcard characters to the query and submit a wildcard search. This is done by using a regex. If the regex is true, it strips out the alphabetical characters on either side of the string and then inititates a wildcard search. This wildcard search does not occur if the query does not match the regex.