Close Kafka connection on SIGINT or SIGTERM, not beforeExit
The beforeExit event occurs when the runtime has drained the event
loop and has no more work to perform, and is about to exit.

However the purpose of this signal handler is to disconnect the
Kafka client which is holding some unsettled promises, which
prevent beforeExit from ever being called.

Instead, handle SIGINT and SIGTERM events directly.