open external links in new tab
This adds a "isExternal" boolean prop to AstroDataLinkWithTooltip and AstroDataLink, which causes links with the prop to open in a new tab. This maintains the default behaviour of opening links in the same tab. The implementation for defining the target prop in AstroDataLink for external links in this commit might not be ideal. Alternatively, a function could be added to check if the link begins with gcn.nasa.gov, which would then change the target in cases where it evaluates as true.