Circulars: enable basic lucene search
this commit adds the most basic form of lucene search, which can be enabled by adding the CIRCULARS_LUCENE feature flag. Additional tuning of field weighting and some type of workaround for the autor search will be required to make it more functional, as well as documentation instructing users on making use of lucene search.