Fix rate limiting due to improperly debounced auto complete
In this context, the `useDebouncedCallback` hook is not working.
Although it delays the execution of the callback until `wait`
seconds of inactivity have passed, it does not decrease the
_number_ of executions --- it just bunches them all together at
once. As a result, we exceed the Cognito rate limit and the page
crashes.

Instead, use the Lodash `debounce` method directly. In my local
sandbox environment, I observe that this approach does result in
properly debounced invocations of the callback.