Gracefully handle build errors in esbuild script
We have two plugins which each take the metafile as an argument.
However, if the build fails, then the metafile arguemnt will not
be set. In this case esbuild will print the original and hopefully
helpful error message, but also unhelpful error messages about the
metafile being undefined. Suppress the latter error messages by
checking first that the metafile is defined.