basic fallback functionality for queries
this ensures that, if a user submits a query that would raise an error with the query_string search method, the existing multi_match method is used as a fallback. The current queryObj constant is somewhat redundant with the addition of this fallback, but this will likely be removed once the feature flag is removed.