update advanced search documentation
this adds more language to the advanced search documentation describing forbidden behaviour that will cause errors when included in queries. This compliments the fallback method of #2930. Additionally, the compound queries section is expanded upon to describe the NOT operator, which can be used in tandem with the AND and OR operators