add basic navigation buttons
change order of buttons

redirect to archive main page if a circular does not exist

Add system for checking for adjacent circulars

This changes the method of finding adjacent circulars so it queries the dynamodb table iteratively until it encounters the next/previous valid circular or it hits a desginated stop point. In the event that no valid circular is found, it will return false, and the button corresponding to that will no longer appear.

remove console log lines

change order of navigation buttons

change to grey out buttons when adjacent circular is not present