Unify interface for Circulars corrections and edits
In the view route for individual Circulars, consolidate the
buttons for requesting a correction and editing a Circular into
a single button. If the user is a moderator, then they get the
option to either submit a correction or save as a new version
on the Circulars edit form itself.

This reduces clutter in the Circulars button bar and places the
decision over what kind of change to make in a more logical
place.