Always show Circulars edit button
Always show the Circulars edit button, even when the user is not
logged in. When the user navigates to either the new Circular page
or the edit page, if they are not signed in then they are
presented with the same dialog box to invite them to sign in and
get a peer endorsement.