Simplify code flow in app client expiration Lambda
Reorganize and simplify the code flow, allowing expiration actions
to occur as we are iterating over the users to get their email
addresses. This should speed up the execution because more
operations are allowed to happen in parallel. However, it should
reduce API rate-limiting because it is less likely that the app
client deletion API calls all happen at once.